test(chaos): decompose large chaos test suites into focused modules
