+++ backend/tests/chaos/test_browser_network_chaos.py
@@ -44,7 +44,6 @@ from app.services.agentic.schemas import (
     SelectorCandidate,
     SelectorDiagnosisReport,
 )
-from app.services.browser.diagnostics import detect_page_state
 from app.services.browser.manager import (
     BrowserManager,
 )
@@ -474,160 +473,4 @@ class TestMidScrapeSessionInvalidation:
             # Verified: loop fast-aborts on first logged_out failure instead of continuing
             assert call_count == 1
             assert len(result.topics_failed) == 1
-            assert result.topics_scraped == 0
-
-
-# ==============================================================================
-# 2. RATE LIMIT & BOT CHALLENGES (CAPTCHA / CLOUDFLARE / ARKOSE) TESTS
-# ==============================================================================
-
-
-class TestRateLimitsAndBotChallenges:
-    """Chaos tests attacking rate limit and bot detection state handling."""
-
-    @pytest.mark.anyio
-    async def test_detect_page_state_rate_limit_variations(self) -> None:
-        """Attack Vector 2A: Validates rate limit text variations."""
-        # 1. Rate limit exceeded banner
-        page_rl = AsyncMock(url="https://x.com/home")
-
-        def loc_rl(selector: str) -> Any:
-            if selector == "text=Rate limit exceeded":
-                return _build_mock_locator(count=1)
-            return _build_mock_locator(count=0)
-
-        page_rl.locator = MagicMock(side_effect=loc_rl)
-        state = await detect_page_state(page_rl)
-        assert state == "rate_limited"
-
-        # 2. Something went wrong banner
-        page_sww = AsyncMock(url="https://x.com/home")
-
-        def loc_sww(selector: str) -> Any:
-            if selector == "text=Something went wrong":
-                return _build_mock_locator(count=1)
-            return _build_mock_locator(count=0)
-
-        page_sww.locator = MagicMock(side_effect=loc_sww)
-        state = await detect_page_state(page_sww)
-        assert state == "rate_limited"
-
-        # 3. Account locked banner
-        page_locked = AsyncMock(url="https://x.com/home")
-
-        def loc_locked(selector: str) -> Any:
-            if selector == "text=Your account has been locked":
-                return _build_mock_locator(count=1)
-            return _build_mock_locator(count=0)
-
-        page_locked.locator = MagicMock(side_effect=loc_locked)
-        state_locked = await detect_page_state(page_locked)
-        assert state_locked == "rate_limited"
-
-    @pytest.mark.anyio
-    async def test_detect_page_state_captcha_sizing_heuristics(self) -> None:
-        """Attack Vector 2B: Differentiates tracking iframes vs visible CAPTCHA walls."""
-        # 1. Hidden / zero-size tracking iframe -> returns 'ok'
-        page_tracking = AsyncMock(url="https://x.com/home")
-        mock_tracking_iframe = AsyncMock()
-        mock_tracking_iframe.is_visible = AsyncMock(return_value=False)
-        mock_tracking_iframe.bounding_box = AsyncMock(return_value=None)
-
-        def loc_tracking(selector: str) -> Any:
-            if selector == "iframe[src*='captcha']":
-                return _build_mock_locator(count=1, all_items=[mock_tracking_iframe])
-            return _build_mock_locator(count=0)
-
-        page_tracking.locator = MagicMock(side_effect=loc_tracking)
-        state_tracking = await detect_page_state(page_tracking)
-        assert state_tracking == "ok"
-
-        # 2. Large visible CAPTCHA modal (> 100x100) -> returns 'captcha'
-        page_captcha = AsyncMock(url="https://x.com/home")
-        mock_captcha_iframe = AsyncMock()
-        mock_captcha_iframe.is_visible = AsyncMock(return_value=True)
-        mock_captcha_iframe.bounding_box = AsyncMock(
-            return_value={"x": 100, "y": 100, "width": 400, "height": 300}
-        )
-
-        def loc_captcha(selector: str) -> Any:
-            if selector == "iframe[src*='captcha']":
-                return _build_mock_locator(count=1, all_items=[mock_captcha_iframe])
-            return _build_mock_locator(count=0)
-
-        page_captcha.locator = MagicMock(side_effect=loc_captcha)
-        state_captcha = await detect_page_state(page_captcha)
-        assert state_captcha == "captcha"
-
-    @pytest.mark.anyio
-    async def test_detect_page_state_arkose_and_cloudflare_blindspots(self) -> None:
-        """Attack Vector 2C: Arkose Labs & Cloudflare challenge detection."""
-        page_arkose = AsyncMock(url="https://x.com/account/access")
-        page_arkose.title = AsyncMock(return_value="Twitter / Arkose Challenge")
-
-        mock_arkose_iframe = AsyncMock()
-        mock_arkose_iframe.is_visible = AsyncMock(return_value=True)
-        mock_arkose_iframe.bounding_box = AsyncMock(
-            return_value={"x": 50, "y": 50, "width": 350, "height": 400}
-        )
-
-        def loc_arkose(selector: str) -> Any:
-            if "arkose" in selector or "arkoselabs" in selector:
-                return _build_mock_locator(count=1, all_items=[mock_arkose_iframe])
-            return _build_mock_locator(count=0)
-
-        page_arkose.locator = MagicMock(side_effect=loc_arkose)
-        state_arkose = await detect_page_state(page_arkose)
-
-        # Verified: Arkose challenge is detected as 'captcha'
-        assert state_arkose == "captcha"
-
-        # Also test Cloudflare Turnstile title check
-        page_cf = AsyncMock(url="https://x.com")
-        page_cf.title = AsyncMock(return_value="Just a moment...")
-        page_cf.locator = MagicMock(return_value=_build_mock_locator(count=0))
-        state_cf = await detect_page_state(page_cf)
-        assert state_cf == "captcha"
-
-    @pytest.mark.anyio
-    async def test_scrape_trending_topics_rate_limited_clean_exit(self) -> None:
-        """Attack Vector 2D: scrape_trending_topics cleanly aborts when home feed is rate limited."""
-        mock_page = AsyncMock()
-        mock_page.url = "https://x.com/home"
-        mock_page.goto = AsyncMock()
-
-        def loc_fn(selector: str) -> Any:
-            if selector == "text=Rate limit exceeded":
-                return _build_mock_locator(count=1)
-            return _build_mock_locator(count=0)
-
-        mock_page.locator = MagicMock(side_effect=loc_fn)
-
-        mock_context = AsyncMock()
-        mock_context.pages = [mock_page]
-
-        mock_manager = MagicMock()
-        mock_manager.get_context = MagicMock(
-            return_value=AsyncMock(
-                __aenter__=AsyncMock(return_value=mock_context),
-                __aexit__=AsyncMock(return_value=None),
-            )
-        )
-
-        with (
-            patch(
-                "scripts.scrape_trending_topics.BrowserManager",
-                return_value=mock_manager,
-            ),
-            patch(
-                "scripts.scrape_trending_topics.random_delay", new_callable=AsyncMock
-            ),
-            patch(
-                "scripts.scrape_trending_topics._resolve_target_user", return_value=None
-            ),
-        ):
-            result = await scrape_trending_topics()
-
-            assert result.status == "rate_limited"
-            assert result.topics_found == 0
             assert result.topics_scraped == 0
